requirements.txt
# Core dependencies
requests>=2.28.0
python-dateutil>=2.8.2
pytz>=2022.1
PyYAML>=6.0

# LinkedIn automation
playwright>=1.30.0

# Email functionality
# Note: smtplib and email are part of Python's standard library

# Content generation via OpenRouter (no SDK required)
# openai and transformers not required for OpenRouter usage

# Web scraping and SEO
beautifulsoup4>=4.11.1

# Utilities
python-dotenv>=0.20.0
scikit-learn>=1.3.0
numpy>=1.24.0